ajout gestion role

+++ backend/src/main/java/com/example/bibliotheque_quali_dev/dto/TopLivreStat.java
@@ -8,7 +8,7 @@ public class TopLivreStat {
     public TopLivreStat() {
     }
 
-    public TopLivreStat(Integer idLivre, String titre, long nbEmprunts) {
+    public TopLivreStat(Integer idLivre, String titre, int nbEmprunts) {
         this.idLivre = idLivre;
         this.titre = titre;
         this.nbEmprunts = nbEmprunts;
